requirements_webapp.txt
annoy
flask
pandas
nltk
numpy
gensim
wget
scikit-learn
gradio
Pillow
requests
torch
torchvision
torchtext